# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"coachee service","description":"HTTP service for handling coaches, coachees and their transactions","version":"1.0"},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/coaches":{"post":{"tags":["coachee"],"summary":"CreateCoach coachee","description":"CreateCoaches creates a base coach","operationId":"coachee#CreateCoach","parameters":[{"name":"CreateCoachRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/CoacheeCreateCoachRequestBody","required":["firstName","lastName","email","phone","tags","description","certifications","programs","introCall"]}}],"responses":{"201":{"description":"Created response.","schema":{"type":"integer","format":"int64"}}},"schemes":["http"]}},"/coaches/{id}":{"post":{"tags":["coachee"],"summary":"UpdateCoach coachee","description":"UpdateCoaches updates a coach","operationId":"coachee#UpdateCoach","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"UpdateCoachRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/CoacheeUpdateCoachRequestBody"}}],"responses":{"202":{"description":"Accepted response."}},"schemes":["http"]}},"/coaches/{id}/availability":{"put":{"tags":["coachee"],"summary":"CreateAvailability coachee","description":"creates an availability for a coach","operationId":"coachee#CreateAvailability","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"CreateAvailabilityRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/CoacheeCreateAvailabilityRequestBody","required":["availability"]}}],"responses":{"202":{"description":"Accepted response."}},"schemes":["http"]}},"/coaches/{id}/availability/{avID}":{"delete":{"tags":["coachee"],"summary":"DeleteAvailability coachee","description":"deletes an availability for a coach","operationId":"coachee#DeleteAvailability","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"avID","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/coaches/{id}/certifications":{"put":{"tags":["coachee"],"summary":"CreateCertification coachee","description":"creates a certification for a coach","operationId":"coachee#CreateCertification","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"CreateCertificationRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/CoacheeCreateCertificationRequestBody","required":["certification"]}}],"responses":{"202":{"description":"Accepted response."}},"schemes":["http"]}},"/coaches/{id}/certifications/{certID}":{"delete":{"tags":["coachee"],"summary":"DeleteCertification coachee","description":"deletes a certification for a coach","operationId":"coachee#DeleteCertification","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"certID","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/coaches/{id}/programs":{"put":{"tags":["coachee"],"summary":"CreateProgram coachee","description":"creates a program for a coach","operationId":"coachee#CreateProgram","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"CreateProgramRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/CoacheeCreateProgramRequestBody","required":["program"]}}],"responses":{"202":{"description":"Accepted response."}},"schemes":["http"]}},"/coaches/{id}/programs/{programID}":{"delete":{"tags":["coachee"],"summary":"DeleteProgram coachee","description":"deletes a program for a coach","operationId":"coachee#DeleteProgram","parameters":[{"name":"id","in":"path","required":true,"type":"integer"},{"name":"programID","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/coaches/{tag}":{"get":{"tags":["coachee"],"summary":"GetCoaches coachee","description":"GetCoaches returns an array of coaches according to a tag and pagination","operationId":"coachee#GetCoaches","parameters":[{"name":"limit","in":"query","required":false,"type":"integer"},{"name":"page","in":"query","required":false,"type":"integer"},{"name":"tag","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/CoacheeGetCoachesResponseBody"}}},"schemes":["http"]}}},"definitions":{"CoacheeCreateAvailabilityRequestBody":{"title":"CoacheeCreateAvailabilityRequestBody","type":"object","properties":{"availability":{"$ref":"#/definitions/availabilityRequestBody"}},"example":{"availability":{"end":765,"id":"Nisi occaecati recusandae voluptatem dolorem omnis.","start":706,"weekDay":4}},"required":["availability"]},"CoacheeCreateCertificationRequestBody":{"title":"CoacheeCreateCertificationRequestBody","type":"object","properties":{"certification":{"$ref":"#/definitions/certificationsRequestBody"}},"example":{"certification":{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023}},"required":["certification"]},"CoacheeCreateCoachRequestBody":{"title":"CoacheeCreateCoachRequestBody","type":"object","properties":{"certifications":{"type":"array","items":{"$ref":"#/definitions/certificationsRequestBody"},"example":[{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023},{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023},{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023}]},"city":{"type":"string","example":"Sint facilis."},"country":{"type":"string","example":"Rerum laborum."},"description":{"type":"string","example":"Quo incidunt ad recusandae quam."},"email":{"type":"string","example":"Est tempora iure id aspernatur."},"firstName":{"type":"string","example":"Et autem voluptatem veniam."},"introCall":{"type":"integer","example":16620491656996686255,"format":"int64"},"lastName":{"type":"string","example":"Inventore fugit distinctio delectus culpa doloremque."},"phone":{"type":"string","example":"Esse molestiae eum et et."},"programs":{"type":"array","items":{"$ref":"#/definitions/programRequestBody"},"example":[{"description":"Hic eveniet placeat reiciendis cum reiciendis.","duration":13130017794459440444,"id":"Enim et quas ex ut.","name":"Aut voluptatem.","sessions":14817489201336631492,"taxPercent":6813321315917995789,"totalPrice":2652402719978733960},{"description":"Hic eveniet placeat reiciendis cum reiciendis.","duration":13130017794459440444,"id":"Enim et quas ex ut.","name":"Aut voluptatem.","sessions":14817489201336631492,"taxPercent":6813321315917995789,"totalPrice":2652402719978733960}]},"tags":{"type":"string","example":"Iusto incidunt odio."}},"example":{"certifications":[{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023},{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023},{"description":"Ex sed et.","id":"Eos nam aperiam.","institution":"Eligendi rerum animi tenetur ipsa debitis molestiae.","month":1,"title":"Quae iure assumenda sequi.","year":2023}],"city":"Quia aliquid numquam.","country":"Dolor sit dolor tempore iusto possimus.","description":"Aliquid saepe velit suscipit delectus fugit.","email":"Aut soluta consectetur.","firstName":"Praesentium minima.","introCall":16087707745518748538,"lastName":"Occaecati quis error consectetur in consequatur labore.","phone":"Tempora fugit et provident dolore provident repellendus.","programs":[{"description":"Hic eveniet placeat reiciendis cum reiciendis.","duration":13130017794459440444,"id":"Enim et quas ex ut.","name":"Aut voluptatem.","sessions":14817489201336631492,"taxPercent":6813321315917995789,"totalPrice":2652402719978733960},{"description":"Hic eveniet placeat reiciendis cum reiciendis.","duration":13130017794459440444,"id":"Enim et quas ex ut.","name":"Aut voluptatem.","sessions":14817489201336631492,"taxPercent":6813321315917995789,"totalPrice":2652402719978733960}],"tags":"Velit eveniet."},"required":["firstName","lastName","email","phone","tags","description","certifications","programs","introCall"]},"CoacheeCreateProgramRequestBody":{"title":"CoacheeCreateProgramRequestBody","type":"object","properties":{"program":{"$ref":"#/definitions/programRequestBody"}},"example":{"program":{"description":"Hic eveniet placeat reiciendis cum reiciendis.","duration":13130017794459440444,"id":"Enim et quas ex ut.","name":"Aut voluptatem.","sessions":14817489201336631492,"taxPercent":6813321315917995789,"totalPrice":2652402719978733960}},"required":["program"]},"CoacheeGetCoachesResponseBody":{"title":"CoacheeGetCoachesResponseBody","type":"object","properties":{"availability":{"type":"array","items":{"$ref":"#/definitions/availabilityResponseBody"},"example":[{"end":580,"id":"Ut repellendus deserunt consectetur accusantium.","start":508,"weekDay":1},{"end":580,"id":"Ut repellendus deserunt consectetur accusantium.","start":508,"weekDay":1}]},"certifications":{"type":"array","items":{"$ref":"#/definitions/certificationsResponseBody"},"example":[{"description":"Deserunt et.","id":"Ipsum sed.","institution":"Natus a facere dicta.","month":5,"title":"Vitae vero qui.","year":2013},{"description":"Deserunt et.","id":"Ipsum sed.","institution":"Natus a facere dicta.","month":5,"title":"Vitae vero qui.","year":2013}]},"city":{"type":"string","example":"Ullam libero officiis in quidem."},"country":{"type":"string","example":"Deserunt eligendi ut quaerat ea quos."},"description":{"type":"string","example":"Ut enim magni cupiditate nihil quam."},"firstName":{"type":"string","example":"Ea atque ipsa."},"id":{"type":"integer","example":14403548211272472435,"format":"int64"},"lastName":{"type":"string","example":"Sint magni mollitia consequatur provident sint."},"pictureURL":{"type":"string","example":"Hic tempora."},"programs":{"type":"array","items":{"$ref":"#/definitions/programResponseBody"},"example":[{"description":"Nesciunt maxime ipsum suscipit mollitia.","duration":4329549923650235340,"id":"Veritatis et.","name":"Deserunt quaerat laboriosam qui.","sessions":11149628812217043023,"taxPercent":1281230411718509234,"totalPrice":16720946541478502871},{"description":"Nesciunt maxime ipsum suscipit mollitia.","duration":4329549923650235340,"id":"Veritatis et.","name":"Deserunt quaerat laboriosam qui.","sessions":11149628812217043023,"taxPercent":1281230411718509234,"totalPrice":16720946541478502871},{"description":"Nesciunt maxime ipsum suscipit mollitia.","duration":4329549923650235340,"id":"Veritatis et.","name":"Deserunt quaerat laboriosam qui.","sessions":11149628812217043023,"taxPercent":1281230411718509234,"totalPrice":16720946541478502871}]},"tags":{"type":"string","example":"Molestias quas iusto facere."}},"example":{"availability":[{"end":580,"id":"Ut repellendus deserunt consectetur accusantium.","start":508,"weekDay":1},{"end":580,"id":"Ut repellendus deserunt consectetur accusantium.","start":508,"weekDay":1},{"end":580,"id":"Ut repellendus deserunt consectetur accusantium.","start":508,"weekDay":1},{"end":580,"id":"Ut repellendus deserunt consectetur accusantium.","start":508,"weekDay":1}],"certifications":[{"description":"Deserunt et.","id":"Ipsum sed.","institution":"Natus a facere dicta.","month":5,"title":"Vitae vero qui.","year":2013},{"description":"Deserunt et.","id":"Ipsum sed.","institution":"Natus a facere dicta.","month":5,"title":"Vitae vero qui.","year":2013}],"city":"Nisi quam.","country":"Provident porro doloremque repellendus nemo.","description":"Excepturi natus nesciunt.","firstName":"Mollitia reiciendis.","id":11671010134317199316,"lastName":"Rerum nulla optio quisquam vitae.","pictureURL":"Fuga saepe sed.","programs":[{"description":"Nesciunt maxime ipsum suscipit mollitia.","duration":4329549923650235340,"id":"Veritatis et.","name":"Deserunt quaerat laboriosam qui.","sessions":11149628812217043023,"taxPercent":1281230411718509234,"totalPrice":16720946541478502871},{"description":"Nesciunt maxime ipsum suscipit mollitia.","duration":4329549923650235340,"id":"Veritatis et.","name":"Deserunt quaerat laboriosam qui.","sessions":11149628812217043023,"taxPercent":1281230411718509234,"totalPrice":16720946541478502871}],"tags":"Facere ipsa voluptate."}},"CoacheeUpdateCoachRequestBody":{"title":"CoacheeUpdateCoachRequestBody","type":"object","properties":{"city":{"type":"string","example":"Rerum rerum officiis quo nobis."},"country":{"type":"string","example":"Atque aliquid consequatur velit voluptatem ab."},"description":{"type":"string","example":"Voluptatem voluptates omnis cumque."},"email":{"type":"string","example":"Fugit amet non asperiores quia non."},"firstName":{"type":"string","example":"Quae tenetur pariatur sunt praesentium aut."},"introCall":{"type":"integer","example":3779802466602698167,"format":"int64"},"lastName":{"type":"string","example":"Ipsum dolores eius ut officia et hic."},"phone":{"type":"string","example":"In qui excepturi vel ut."},"pictureURL":{"type":"string","example":"Vitae id fugit delectus quia possimus."},"stripeID":{"type":"string","example":"Neque sunt voluptatem."},"tags":{"type":"string","example":"Inventore voluptatibus odio minima."}},"example":{"city":"Dolorum reprehenderit exercitationem est dolores.","country":"Eius corporis voluptas sapiente deserunt molestiae.","description":"Dolorem quo.","email":"Sint qui odit quia voluptas labore voluptatum.","firstName":"Tempore accusantium porro nisi aut.","introCall":16773765413436263084,"lastName":"Beatae non quia blanditiis similique.","phone":"Assumenda distinctio.","pictureURL":"Iure vel officia non et voluptates.","stripeID":"Laboriosam eveniet animi est et suscipit.","tags":"Neque quis sunt esse placeat consequatur."}},"availabilityRequestBody":{"title":"availabilityRequestBody","type":"object","properties":{"end":{"type":"integer","example":160,"minimum":0,"maximum":1440},"id":{"type":"string","example":"Illum dolore amet fugiat ab dolorum."},"start":{"type":"integer","example":375,"minimum":0,"maximum":1440},"weekDay":{"type":"integer","example":5,"minimum":0,"maximum":6}},"description":"represents a coach availability","example":{"end":870,"id":"Culpa quo repellat.","start":141,"weekDay":0}},"availabilityResponseBody":{"title":"availabilityResponseBody","type":"object","properties":{"end":{"type":"integer","example":908,"minimum":0,"maximum":1440},"id":{"type":"string","example":"Quia sit est enim unde illo dolorem."},"start":{"type":"integer","example":382,"minimum":0,"maximum":1440},"weekDay":{"type":"integer","example":5,"minimum":0,"maximum":6}},"description":"represents a coach availability","example":{"end":1269,"id":"Qui quidem sint et voluptates tenetur in.","start":1383,"weekDay":4}},"certificationsRequestBody":{"title":"certificationsRequestBody","type":"object","properties":{"description":{"type":"string","example":"Repellat id aperiam fugit facere repellat."},"id":{"type":"string","example":"Quae magnam ut vero autem magnam rerum."},"institution":{"type":"string","example":"Architecto amet sint facere est incidunt rerum."},"month":{"type":"integer","example":5,"minimum":1,"maximum":12},"title":{"type":"string","example":"Autem harum aut beatae."},"year":{"type":"integer","example":2008,"minimum":1900,"maximum":2100}},"description":"represents a coach certification","example":{"description":"Enim saepe iste assumenda debitis ut qui.","id":"Accusamus autem.","institution":"Id eveniet.","month":3,"title":"Quo sit sit accusantium neque.","year":1905},"required":["title","description","institution","month","year"]},"certificationsResponseBody":{"title":"certificationsResponseBody","type":"object","properties":{"description":{"type":"string","example":"Dicta eligendi blanditiis vero velit ea."},"id":{"type":"string","example":"Ipsam repellat maxime est."},"institution":{"type":"string","example":"Illum rem."},"month":{"type":"integer","example":8,"minimum":1,"maximum":12},"title":{"type":"string","example":"Sapiente est dolores occaecati inventore aliquid."},"year":{"type":"integer","example":1925,"minimum":1900,"maximum":2100}},"description":"represents a coach certification","example":{"description":"Explicabo ratione qui minima quis hic sit.","id":"Iusto quia cum officia soluta quo deserunt.","institution":"Ipsam similique eum iusto.","month":10,"title":"Voluptatem ipsa quas et fuga quod suscipit.","year":2055},"required":["title","description","institution","month","year"]},"programRequestBody":{"title":"programRequestBody","type":"object","properties":{"description":{"type":"string","example":"Sed eum veritatis beatae."},"duration":{"type":"integer","example":664489634866332272,"format":"int64"},"id":{"type":"string","example":"Est qui laborum qui."},"name":{"type":"string","example":"Quia totam quia minima doloremque necessitatibus pariatur."},"sessions":{"type":"integer","example":13156642908273162186,"format":"int64"},"taxPercent":{"type":"integer","example":16937435541461840426,"format":"int64"},"totalPrice":{"type":"integer","example":5305864060907313485,"format":"int64"}},"description":"represents a coach's programs","example":{"description":"Esse sit aliquam.","duration":1012218529499208753,"id":"Quos vitae.","name":"Soluta repellat unde aut expedita vero.","sessions":6971254298652483444,"taxPercent":6784042706580477644,"totalPrice":4758501574718213800},"required":["name","sessions","duration","description","totalPrice","taxPercent"]},"programResponseBody":{"title":"programResponseBody","type":"object","properties":{"description":{"type":"string","example":"Eligendi et."},"duration":{"type":"integer","example":12473513830421858177,"format":"int64"},"id":{"type":"string","example":"Id deserunt natus omnis debitis omnis."},"name":{"type":"string","example":"Aperiam quia qui facilis dolorem."},"sessions":{"type":"integer","example":13287496684956567465,"format":"int64"},"taxPercent":{"type":"integer","example":12051872438917094982,"format":"int64"},"totalPrice":{"type":"integer","example":451748246373320723,"format":"int64"}},"description":"represents a coach's programs","example":{"description":"Aut veritatis aut.","duration":6983943200034854534,"id":"Dolor sunt eos.","name":"Quia quia blanditiis.","sessions":1284093384083185755,"taxPercent":8497091778453373750,"totalPrice":13950912840516961539},"required":["name","sessions","duration","description","totalPrice","taxPercent"]}}}